refactor: Move clip_tokens function from pr_processing to utils module, and add tests
