+++ pr_agent/git_providers/git_provider.py
@@ -63,7 +63,7 @@ class GitProvider(ABC):
 
     def get_pr_description(self, *, full: bool = True) -> str:
         from pr_agent.config_loader import get_settings
-        from pr_agent.algo.pr_processing import clip_tokens
+        from pr_agent.algo.utils import clip_tokens
         max_tokens_description = get_settings().get("CONFIG.MAX_DESCRIPTION_TOKENS", None)
         description = self.get_pr_description_full() if full else self.get_user_description()
         if max_tokens_description:
